sep button, old Ediable->EditName
