+++ zenqt/ui/editor/button.py
@@ -1,0 +1,90 @@
+from . import *
+
+from ..utils import setKeepAspect
+
+
+class QDMGraphicsButton(QGraphicsItem):
+    def __init__(self, parent):
+        super().__init__(parent)
+
+        self.node = parent
+        self.name = None
+
+        self.initLabel()
+        self.setChecked(False)
+
+    def initLabel(self):
+        self.label = QGraphicsTextItem(self)
+        self.label.setPos(0, - TEXT_HEIGHT * 0.1)
+        font = QFont()
+        font.setPointSize(style['socket_text_size'])
+        self.label.setFont(font)
+
+        document = self.label.document()
+        option = document.defaultTextOption()
+        option.setAlignment(Qt.AlignCenter)
+        document.setDefaultTextOption(option)
+
+    def setText(self, name):
+        self.name = name
+        self.label.setPlainText(name)
+
+    def getCircleBounds(self):
+        return (0, 0, self._width, self._height)
+
+    def boundingRect(self):
+        return QRectF(*self.getCircleBounds()).normalized()
+
+    def paint(self, painter, styleOptions, widget=None):
+        button_color = 'button_selected_color' if self.checked else 'button_color'
+        painter.fillRect(*self.getCircleBounds(), QColor(style[button_color]))
+
+    def setChecked(self, checked):
+        self.checked = checked
+
+        text_color = 'button_selected_text_color' if self.checked else 'button_text_color'
+        self.label.setDefaultTextColor(QColor(style[text_color]))
+
+    def setWidthHeight(self, width, height):
+        self._width = width
+        self._height = height
+        self.label.setTextWidth(width)
+
+    def on_click(self):
+        self.setChecked(not self.checked)
+
+    def mousePressEvent(self, event):
+        super().mousePressEvent(event)
+        self.on_click()
+
+    def setGeometry(self, rect):
+        x = rect.x()
+        y = rect.y()
+        w = rect.width()
+        h = rect.height()
+        self.setPos(x, y)
+        self.setWidthHeight(w, h)
+
+
+class QDMGraphicsCollapseButton(QGraphicsSvgItem):
+    def __init__(self, parent):
+        super().__init__(parent)
+        self.node = parent
+
+        self._renderer = QSvgRenderer(asset_path('unfold.svg'))
+        self.update_svg(False)
+
+    def update_svg(self, collapsed):
+        svg_filename = ('collapse' if collapsed else 'unfold') + '.svg'
+        self._renderer.load(asset_path(svg_filename))
+        setKeepAspect(self._renderer)
+        self.setSharedRenderer(self._renderer)
+
+    def mousePressEvent(self, event):
+        super().mouseMoveEvent(event)
+        self.node.collapsed = not self.node.collapsed
+        if self.node.collapsed:
+            self.node.collapse()
+        else:
+            self.node.unfold()
+
