+++ zenqt/ui/editor/makedict.py
@@ -1,5 +1,6 @@
 from . import *
 
+
 class QDMGraphicsSocketEditName(QDMGraphicsSocket):
     def __init__(self, parent=None):
         super().__init__(parent)
